+++ CHANGELOG.md
@@ -2,10 +2,14 @@ ThreePhasePowerModels.jl Change Log
 ===================================
 
 ### staged
+- none
+
+### v0.3.1
 - Allow for arbitrarily named sourcebus
 - Add json parser
 - Add support for additional load models (constant power, constant impedance, constant current; delta or wye connected) (#127)
 - Fix bug in OpenDSS parse of Capacitors [zbase factor and wrong sign] (#138)
+- Add voltage balance constraints (#129)
 
 ### v0.3.0
 - Update to JuMP v0.19/MathOptInterface